requirements.txt
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
click-plugins==1.1.1
cligj==0.7.2
cycler==0.11.0
Fiona==1.8.21
fonttools==4.29.1
geopandas==0.10.2
idna==3.3
kiwisolver==1.3.2
matplotlib==3.5.1
munch==2.5.0
numpy==1.22.2
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
pygeos==0.12.0
pyparsing==3.0.7
pyproj==3.3.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2021.3
requests==2.27.1
Shapely==1.8.1.post1
six==1.16.0
urllib3==1.26.8
